app/requirements.txt
python-decouple
djangorestframework